Restructure and add the unit tests.
