+++ tests/ContaoManagerBundleTest.php
@@ -14,11 +14,24 @@ namespace Contao\ManagerBundle\Tests;
 
 use Contao\ManagerBundle\ContaoManagerBundle;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 class ContaoManagerBundleTest extends TestCase
 {
     public function testInstantiation(): void
     {
         $this->assertInstanceOf('Contao\ManagerBundle\ContaoManagerBundle', new ContaoManagerBundle());
+    }
+
+    public function testBuild(): void
+    {
+        $container = new ContainerBuilder();
+
+        $bundle = new ContaoManagerBundle();
+        $bundle->build($container);
+
+        $passes = $container->getCompilerPassConfig()->getBeforeOptimizationPasses();
+
+        $this->assertInstanceOf('Contao\ManagerBundle\DependencyInjection\Compiler\SwiftMailerPass', $passes[3]);
     }
 }
